+++ app/controllers/command_controller.py
@@ -1,21 +1,20 @@
 from litestar import Litestar, get
-from litestar.contrib.sqlalchemy.base import UUIDBase, UUIDAuditBase
-from app.models.user_model import UserModel
-from app.models.post_model import Post
-from app.models.comment_model import Comment
-from app.models.like_dislike_model import LikeDislike
-from app.models.tag_model import Tag
-from app.models.event_model import Event
+from litestar.contrib.sqlalchemy.base import UUIDAuditBase, UUIDBase
 from litestar.contrib.sqlalchemy.plugins.init import SQLAlchemyInitPlugin
 from litestar.contrib.sqlalchemy.plugins.init.config import (
     AsyncSessionConfig,
     SQLAlchemyAsyncConfig,
 )
-from app.utils.pika import RabbitMQConnection
 
-
-from app.controllers.command.UserController import UserController
 from app.controllers.command.TagController import TagController
+from app.controllers.command.UserController import UserController
+from app.models.comment_model import Comment
+from app.models.event_model import Event
+from app.models.like_dislike_model import LikeDislike
+from app.models.post_model import Post
+from app.models.tag_model import Tag
+from app.models.user_model import User
+from app.utils.pika import RabbitMQConnection
 
 
 @get("/")
@@ -41,6 +40,7 @@ async def on_startup(app: Litestar) -> None:
     async with app.state.db_engine.begin() as conn:
         # await conn.run_sync(UUIDBase.metadata.drop_all) # <---- Uncomment this if you need to reset database
         await conn.run_sync(UUIDBase.metadata.create_all)
+
 
 app = Litestar(
     plugins=[SQLAlchemyInitPlugin(sqlalchemy_config)],