+++ app/models/like_dislike_model.py
@@ -1,13 +1,14 @@
-
 from typing import Annotated
 
 from litestar.contrib.sqlalchemy.base import UUIDAuditBase, UUIDBase
 from litestar.contrib.sqlalchemy.dto import SQLAlchemyDTO
 from litestar.contrib.sqlalchemy.repository import SQLAlchemyAsyncRepository
 from litestar.dto import DTOConfig
-from sqlalchemy import ForeignKey, Table, Column
+from sqlalchemy import Column, ForeignKey, Table
 from sqlalchemy.orm import Mapped, mapped_column, relationship
+
 from app.utils.controller import Service
+
 from .post_model import posts_shared_association
 
 
@@ -15,7 +16,7 @@ class LikeDislike(UUIDBase):
     review_type: Mapped[bool]
 
     reviewed_by_id = Column(ForeignKey('users.id'))
-    reviewed_by = relationship("UserModel", back_populates="likes_dislikes")
+    reviewed_by = relationship("User", back_populates="likes_dislikes")
 
     reviewed_on_id = Column(ForeignKey('posts.id'))
     reviewed_on = relationship("Post", back_populates="likes_dislikes")