+++ app/utils/mongo/collections.py
@@ -1,0 +1,43 @@
+from pymongo.collection import Collection
+
+from .mongo_connect import mongo_db
+
+
+class CollectionController:
+    def __init__(self, collection):
+        self.collection: Collection = collection
+
+    def add_document(self, document):
+        return self.collection.insert_one(document)
+
+    def update_document(self, document_id, update_data):
+        filter_query = {"_id": document_id}
+        update_query = {"$set": update_data}
+        return self.collection.update_one(filter_query, update_query)
+
+    def remove_document(self, document_id):
+        filter_query = {"_id": document_id}
+        return self.collection.delete_one(filter_query)
+
+    def find_document(self, query):
+        return self.collection.find_one(query)
+
+    def find_documents(self, query):
+        return list(self.collection.find(query))
+
+
+user_collection = CollectionController(mongo_db['users'])
+post_collection = CollectionController(mongo_db['posts'])
+comment_collection = CollectionController(mongo_db['comments'])
+tag_collection = CollectionController(mongo_db['tags'])
+event_collection = CollectionController(mongo_db['events'])
+like_dislike_collection = CollectionController(mongo_db['like_dislike'])  # ???
+
+mongo_collections = {
+    "users": user_collection,
+    "posts": post_collection,
+    "comments": comment_collection,
+    "tags": tag_collection,
+    "events": event_collection,
+    "like_dislike": like_dislike_collection,
+}