+++ CHANGELOG.md
@@ -1,2 +1,6 @@
-22.10.2023
-- [x] Initialized repository with basic boilerplate structure
+DP 16.11.2023
+- Altered User model
+- Refactored mongo infrastructure
+
+DP 22.10.2023
+- Initialized repository with basic boilerplate structure